Fix custom theme
