+++ wormhole-connect/src/components/Header.tsx
@@ -5,12 +5,12 @@ export type Alignment = 'center' | 'left' | 'right';
 
 type StyleProps = { align: Alignment; fontSize: number };
 const useStyles = makeStyles<StyleProps>()(
-  (theme: any, { align, fontSize }) => ({
+  (theme, { align, fontSize }) => ({
     title: {
       fontSize: `${fontSize}px`,
       width: '100%',
       textAlign: align,
-      fontFamily: theme.palette.font.header,
+      fontFamily: theme.typography.fontFamily,
       [theme.breakpoints.down('sm')]: {
         fontSize: '24px',
       },
