+++ wormhole-connect/src/createPalette.d.ts
@@ -1,0 +1,55 @@
+import '@mui/material/styles';
+
+declare module '@mui/material/styles' {
+  interface Palette {
+    button: {
+      primary: string;
+      primaryText: string;
+      disabled: string;
+      disabledText: string;
+      action: string;
+      actionText: string;
+      hover: string;
+    };
+    options: {
+      hover: string;
+      select: string;
+    };
+    card: {
+      background: string;
+      secondary: string;
+      elevation: string;
+    };
+    popover: {
+      background: string;
+      secondary: string;
+      elevation: string;
+    };
+  }
+
+  interface PaletteOptions {
+    button?: {
+      primary: string;
+      primaryText: string;
+      disabled: string;
+      disabledText: string;
+      action: string;
+      actionText: string;
+      hover: string;
+    };
+    options?: {
+      hover: string;
+      select: string;
+    };
+    card?: {
+      background: string;
+      secondary: string;
+      elevation: string;
+    };
+    popover?: {
+      background: string;
+      secondary: string;
+      elevation: string;
+    };
+  }
+}
